// src/features/reports/types/report-summary.ts
export type ReportSummary = {
  totalSchools: number;
  totalQrCodes: number;
  assignedQrCodes: number;
  usedQrCodes: number;
  archivedQrCodes: number;
  totalVrRecords: number;
  totalQrRegistrations: number;
  totalMatches: number;
  unmatchedVrRecords: number;
  unmatchedQrRegistrations: number;
  attendedRegistrations: number;
  notAttendedRegistrations: number;
  courseEnrollments: number;
};